feat: release notes in-app + update-state fixtures; CI clippy green
- release notes: pick_update carries the GitHub release body; staged
  updates are a proper StagedUpdate{version, staged, notes}; Settings
  shows the notes (markdown-lite: #-headings render medium-weight,
  capped at 14 lines) under the Restart button, BEFORE the user commits
  to restarting
- progress phases: manual check reports "downloading X…" between the
  cheap check and the long download instead of one silent spinner
- sidebar hint shortened to "update ready · X" (was clipping)
- fixture states settings-update-ready / settings-update-checking pose
  both surfaces with zero network; checking fixture caught a branch-
  order bug (dev-disabled button shadowed the in-flight spinner)
- clippy -D warnings clean: stepper arity allow (established builder
  shape), ProvenanceSides alias, items-before-tests ordering, test
  struct-literal init, redundant cast

+++ crates/app/src/perf.rs
@@ -44,7 +44,7 @@ pub fn time<T>(tag: &'static str, f: impl FnOnce() -> T) -> T {
             eprintln!(
                 "perf[{tag}] n={}/s avg={}µs max={}µs",
                 agg.n,
-                (agg.sum.as_micros() as u64) / agg.n.max(1) as u64,
+                (agg.sum.as_micros() as u64) / agg.n.max(1),
                 agg.max.as_micros(),
             );
             agg.n = 0;
